fix: compile fixes for v0.5 rebrand and feature completion
- Fix runtime.rs: remove extra closing brace in handle_gossip_message, add wildcard match arm for non-exhaustive JobMessage match
- Fix lib.rs: add missing pub mod run declaration
- Fix proxy/mod.rs: remove 2 duplicate is_hop_by_hop functions, remove unsupported KEEP_ALIVE header constant
- Fix proxy/payment.rs: fix borrow checker issue (clone before drop)
- Fix inference/distributed.rs: fix generate import to chat_completion, fix type annotation for oneshot channel
- Fix inference/distribution.rs: fix bandwidth_mbps type (f64 -> u32)
- Fix executor/mod.rs + executor/remote.rs: add missing ttfb_ms field to InferenceResult constructors
- Fix agent/runtime.rs: add missing hitl_store None argument to Self::new call
- Fix tools/mod.rs: re-export ApprovalRequirement
- Fix cli/serve.rs: add missing hitl_store field to WebState constructor
- Fix db/mod.rs: add MsgPack error variant with From impl for rmp_serde::decode::Error
- Fix tests/examples_contract.rs: update crate name bkg-peer -> bkg_p2p
- Run cargo fmt on all source files

+++ src/inference/distribution.rs
@@ -469,7 +469,11 @@ impl ModelDistributor {
     /// Generate announcements for all available models.
     pub async fn create_announcements(&self, local_peer_id: &str) -> Vec<ModelAnnouncement> {
         // Calculate total available bandwidth from all active downloads
-        let total_bandwidth_bps = self.pending_downloads.read().await.iter()
+        let total_bandwidth_bps = self
+            .pending_downloads
+            .read()
+            .await
+            .iter()
             .filter_map(|(_, state)| {
                 let speed_bps = state.download_speed();
                 if speed_bps > 0.0 {
@@ -479,9 +483,9 @@ impl ModelDistributor {
                 }
             })
             .sum::<f64>();
-        
+
         let bandwidth_mbps = if total_bandwidth_bps > 0.0 {
-            Some((total_bandwidth_bps / 1_000_000.0) as f64)
+            Some((total_bandwidth_bps / 1_000_000.0) as u32)
         } else {
             None
         };
